\name{sim_nqc_data}
\alias{sim_nqc_data}
\title{
Artificial data, that imitates non-quality controlled data
}
\description{
The data set 'sim_nqc_data' contans 100 observations and 14 variables with some not well formatted and missing values. The data is complete artificial and only intended as an application example for the package.
}
\usage{
sim_nqc_data
}
\format{
A data frame with 100 observations on 11 (character) variables.
}
\author{
Andreas Schulz
}

\examples{

head(sim_nqc_data)

}
\keyword{data}
