\name{vtype-package}
\alias{vtype-package}
\docType{package}
\title{
Estimates the Variable Type in Error Afflicted Data.
}
\description{
Estimates the type of variables in non-quality controlled data. The prediction is based on pre-trained random forest model, trained on over 5000 medical variables with OOB accuracy of 9999pct. The accuracy can hardy depend on type and coding style of data.
}
\details{
\tabular{ll}{
Package: \tab vtype\cr
Type: \tab Package\cr
Version: \tab 0.8\cr
Date: \tab 2021-05-13\cr
License: \tab GPL version 3 or newer \cr
}
}
\author{
Andreas Schulz, PhD
Maintainer: Andreas Schulz <ades-s@web.de>
}
\keyword{type}
\keyword{variable}
\keyword{prediction}
\keyword{data}
